requirements.txt
APScheduler==3.7.0
asgiref==3.3.1
atomicwrites==1.4.0
attrs==20.2.0
Automat==20.2.0
bcrypt==3.2.0
bytestring-splitter==2.2.0
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
click==7.1.2
colorama==0.4.4
ConfigArgParse==1.2.3
constant-sorrow==0.1.0a9
constantly==15.1.0
cryptography==3.2.1
decorator==4.4.2
Django==3.1.3
django-apscheduler==0.5.2
django-cors-headers==3.5.0
django-crontab==0.7.1
et-xmlfile==1.0.1
Flask==1.1.2
Flask-BasicAuth==0.2.0
gevent==20.12.1
geventhttpclient==1.4.5
greenlet==0.4.17
hyperlink==20.0.1
idna==2.10
importlib-metadata==3.3.0
incremental==17.5.0
iniconfig==1.1.1
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.11.2
jsonpath-rw==1.4.0
locust==1.4.1
MarkupSafe==1.1.1
msgpack==1.0.2
msgpack-python==0.5.6
multi-key-dict==2.0.3
mysqlclient @ file:///C:/Users/%E4%B8%AD%E6%8B%93%E7%94%B5%E5%95%86tcy/PycharmProjects/all_test/mysqlclient-1.4.6-cp37-cp37m-win_amd64.whl
ntlm-auth==1.5.0
numpy==1.19.2
openpyxl==3.0.5
packaging==20.8
paramiko==2.7.2
pbr==5.5.1
pluggy==0.13.1
ply==3.11
psutil==5.8.0
py==1.10.0
pycparser==2.20
PyHamcrest==2.0.2
PyMySQL==0.10.1
PyNaCl==1.4.0
pyOpenSSL==20.0.0
pyparsing==2.4.7
pytest==6.2.1
python-jenkins==1.7.0
pytz==2020.4
pywin32==300
pywinrm==0.4.1
pyzmq==20.0.0
requests==2.24.0
requests-ntlm==1.1.0
simplejson==3.17.2
six==1.15.0
SQLAlchemy==1.3.22
sqlparse==0.4.1
toml==0.10.2
Twisted==20.3.0
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.25.11
Werkzeug==1.0.1
xlrd==1.2.0
xlutils==2.0.0
xlwt==1.3.0
xmltodict==0.12.0
zipp==3.4.0
zope.event==4.5.0
zope.interface==5.1.2
